+++ app/composables/vatsim/pilots.ts
@@ -14,10 +14,14 @@ import type { Pixel } from 'ol/pixel.js';
 import { isHideMapObject } from '~/composables/settings';
 import { collapsingWithOverlay } from '~/composables';
 
-export function usePilotRating(pilot: VatsimShortenedAircraft, short = false): string[] {
+export function usePilotRating(pilot: VatsimShortenedAircraft, short = false, noneIfDefault = false): string[] {
     const dataStore = useDataStore();
 
-    const ratings: string[] = [dataStore.vatsim.data.pilot_ratings.value.find(x => x.id === pilot.pilot_rating)?.[short ? 'short_name' : 'long_name'] ?? ''];
+    const ratings: string[] = [];
+    if (!noneIfDefault || pilot.pilot_rating) {
+        const rating = dataStore.vatsim.data.pilot_ratings.value.find(x => x.id === pilot.pilot_rating)?.[short ? 'short_name' : 'long_name'];
+        if (rating) ratings.push(rating);
+    }
     if (pilot.military_rating) ratings.push(dataStore.vatsim.data.military_ratings.value.find(x => x.id === pilot.pilot_rating)?.[short ? 'short_name' : 'long_name'] ?? pilot.military_rating.toString());
 
     return ratings;