+++ CHANGELOG.md
@@ -5,11 +5,18 @@
 ## Alpha changes
 
 - Improved VATGlasses update speed
+- Fixed labels not rendering for tracked pilots and own flight
+- Moved booked until to same line with time online
+- Redesigned time online
 
 ## Improvements and changes
 
-- Significantly optimized basic layer in terms of performance and memory usage, at a cost of low quality on high zoom
+- Significantly optimized basic layer in terms of performance and memory usage, at a cost of low quality on high zoom levels
 - Removed Sentry to improve bundle size. It's not like I was checking errors anyway
+
+## Bug Fixes
+
+- Removed duplication logic when controller is rendered in VATGlasses
 
 # [2.0.0-alpha.9.1]