+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+# [2.0.0-alpha.9.2]
+
+## Alpha changes
+
+- Improved VATGlasses update speed
+- Fixed labels not rendering for tracked pilots and own flight
+- Moved booked until to same line with time online
+- Redesigned time online
+
+## Improvements and changes
+
+- Significantly optimized basic layer in terms of performance and memory usage, at a cost of low quality on high zoom levels
+- Removed Sentry to improve bundle size. It's not like I was checking errors anyway
+
+## Bug Fixes
+
+- Removed duplication logic when controller is rendered in VATGlasses
+
 # [2.0.0-alpha.9.1]
 
 ## Alpha changes
@@ -11,7 +29,7 @@
 - Improved vertical speed update
 - Implemented dynamic flight plan parameters column count
 
-## Improvements and changed
+## Improvements and changes
 
 - Added an ability to zoom to any controller, including CTR
 - Controllers will now be focused when opened from search - same as with aircraft and airports